pyproject.toml
[tool.poetry]
name = "loop-dune"
version = "0.1.0"
description = "Collect and analyze Loop data"
authors = ["Rangonomics <rangonomics@gmail.com>"]
packages = [
    { include = "loop_dune" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,!=3.9.7"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
eth-typing = "^4.0.0"
tqdm = "^4.66.2"
colorama = "^0.4.6"
requests = "^2.31.0"
streamlit = "^1.32.0"
plotly = "^5.19.0"
schedule = "^1.2.1"
web3 = "6.10.0"
setuptools = "^78.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
PyYAML = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
loop-collect = "loop_dune.collector:main"
loop-upload = "loop_dune.dune_uploader:main"
loop-dashboard = "loop_dune.dashboard:main"
loop-sync = "loop_dune.sync:main"